[Update] Map Conponent
