+++ Prefabs/FX_Healing_Cirle_02.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8403ce52fd82f426a866dcb97593552d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
