+++ Prefabs/FX_Ritual_Circle_01.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8fad6460f54d842a4a5f582b39130054
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
