+++ Scripts/FeedBacks/PowerUpEffect.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 45f505ea31f5e48dc807b6b3bb9712ef
+guid: c5003b07a85b9451cb2443e476bb7e10
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
