+++ Scripts/Map/Location.cs
@@ -7,8 +7,13 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+/// <summary>
+/// 緯度と経度を情報にもったデータクラス
+/// </summary>
 public class Location
 {
+    //経度
     public double Latitude;
+    //緯度
     public double Longitude;
 }
