+++ Scripts/Map/LonLatGetter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e97e86799a86040aa957299eb9e58fd7
+guid: 45f505ea31f5e48dc807b6b3bb9712ef
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
